services/oltp/requirements.txt
Flask==2.1.2
pydantic==1.9.2
kafka-python==2.0.2
gevent==21.12.0
flasgger==0.9.5
pydantic[dotenv]==1.9.2
Flask-Pydantic==0.10.0
gunicorn==20.1.0
Flask-JWT-Extended==4.4.4
pytest==7.1.2
python-logstash==0.4.8
sentry-sdk==1.9.9
blinker==1.5
pymongo==4.2.0
flask-pydantic-spec==0.3.1
Werkzeug==2.1.2
motor==3.0.0
